python/src/requirements.txt
Adafruit-DHT==1.4.0
dht11==0.1.0
RPi.GPIO==0.7.0
